Codewars: add 6_kyu_multiples_of_3_or_5
If we list all the natural numbers below 10 that are multiples of 3 or
5, we get 3, 5, 6 and 9. The sum of these multiples is 23.

Finish the solution so that it returns the sum of all the multiples of 3
or 5 below the number passed in. Additionally, if the number is
negative, return 0 (for languages that do have them).

Note: If the number is a multiple of both 3 and 5, only count it once.

Courtesy of projecteuler.net (Problem 1)

Link: https://projecteuler.net/problem=1
Link: https://www.codewars.com/kata/514b92a657cdc65150000006
Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>