Codewars: add 7_kyu_beginner_series_3_sum_of_numbers
Given two integers a and b, which can be positive or negative, find the
sum of all the integers between and including them and return it. If the
two numbers are equal return a or b.

Note: a and b are not ordered!
Examples

GetSum(1, 0) == 1   // 1 + 0 = 1
GetSum(1, 2) == 3   // 1 + 2 = 3
GetSum(0, 1) == 1   // 0 + 1 = 1
GetSum(1, 1) == 1   // 1 Since both are same
GetSum(-1, 0) == -1 // -1 + 0 = -1
GetSum(-1, 2) == 2  // -1 + 0 + 1 + 2 = 2

Link: https://www.codewars.com/kata/55f2b110f61eb01779000053
Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>