Codewars: add 5_kyu_can_you_get_the_loop
You are given a node that is the beginning of a linked list. This list
always contains a tail and a loop. Your objective is to determine the
length of the loop.

For example in the following picture the tail's size is 3 and the loop
size is 12:

// Use the `next` field to get the following node.
Node* nextNode = nodePtr->next;

Note: do NOT mutate the nodes!

Link: https://www.codewars.com/kata/52a89c2ea8ddc5547a000863
Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>