Codewars/6_kyu_train_your_skills_in_creation_of_classes: create README.txt
In this kata you train your skills in creation of
Classes.
Handled topics are:
    - constructor,
    - data members,
    - class functions,
    - copy constructor,
    - operator overloading.

The task is simple.
With a help of sample tests you are creating a X()
class.  X class contains two data members m_a and m_b
(default values are [1,2].)
After all missing components of the class are
written, it is expected  that following lines of code
will be compiled (see Initial Solution).
Have fun on training!

Note:

You'll have to implement all the mentioned in
"Expected" user cases of the class (see Initial
Solution) of the class. Until there, you will see
compiler errors. Don't forget to select how are you
satisfied with this kata and/or leave your proposals
or comments.

Link:
https://www.codewars.com/kata/5ab4f002379d20e82500008c

Signed-off-by: Alwan Adiuntoro <alwanadiuntoro@gmail.com>